refactor: app wave 9 slice F test support densify (#4117)
Consolidate units-panel scenario builders and expansion helpers into
units_panel_test_shared, extract ct_resource_cell harness helpers, merge
panel support smoke tests outside the support LOC tree, and ratchet
appTestSupportLocCeiling from 8720 to 8302.
