+++ SPEC/program/repo-lint.md
@@ -234,7 +234,7 @@
 | `tool/check_app_e2e_support_lib_file_size.dart` | PR-blocking physical-line gate: `packages/colonizethis_app_e2e_support/lib/**/*.dart` must stay at or below **700** lines; oversized files live on a **shrink-only** allowlist (`appE2eSupportLibFileSizeAllowlistForTests`) — missing or now-under-cap entries fail as stale (Refs #4075); rule `repo.app_e2e_support_lib_file_size`. |
 | `tool/check_app_e2e_support_test_file_size.dart` | PR-blocking physical-line gate: `packages/colonizethis_app_e2e_support/test/**/*.dart` must stay at or below **800** lines; oversized files live on a **shrink-only** allowlist (`appE2eSupportTestFileSizeAllowlistForTests`) — missing or now-under-cap entries fail as stale (Refs #4075); rule `repo.app_e2e_support_test_file_size`. Harness modules under `test/support/` should prefer ≤600 physical lines. |
 | `tool/check_app_test_support_file_size.dart` | PR-blocking physical-line gate: `app/test/support/**/*.dart` must stay at or below **600** lines; oversized files live on a **shrink-only** allowlist (`appTestSupportFileSizeAllowlistForTests`) — missing or now-under-cap entries fail as stale (Refs #4048); rule `repo.app_test_support_file_size`. Distinct from `repo.app_test_file_size` (800) and from `repo.ai_s7d_support_suite_size` (600, no allowlist). |
-| `tool/check_app_test_support_loc.dart` | Ratchet ceiling for all Dart under `app/test/support/` (physical LOC; post-#4117 slice F ceiling **8720**, measured ≈8720); fail on net regrowth (Refs #4021, #4035, #4048, #4058, #4117); rule `repo.app_test_support_loc` |
+| `tool/check_app_test_support_loc.dart` | Ratchet ceiling for all Dart under `app/test/support/` (physical LOC; post-#4117 slice F ceiling **8302**, measured ≈8302); fail on net regrowth (Refs #4021, #4035, #4048, #4058, #4117); rule `repo.app_test_support_loc` |
 | `tool/check_app_test_no_duplicate_dark_token_asserts.dart` | AST gate: forbid private top-level `_expectMuted*` / `_expectEditorialMonocle*` helpers outside `app/test/support/` — suites must call shared APIs in `editorial_monocle_dark_token_assertions.dart` (Refs #4013); rule `repo.app_test_no_duplicate_dark_token_asserts` |
 | `tool/check_app_test_no_duplicate_scaffolding.dart` | Full-tree AST gate: forbid reintroduced min-viewport / Widgetbook use-case / trade-screen host / units-panel `Game(` scaffolding, plus naval-part `_pumpNaval`/`pumpNaval` clones, military-suite `_pumpMilitary`/`pumpMilitary` clones, technology-suite `_pumpPanel`/`pumpPanel` clones, `Scaffold(body: Center(...))` dialog-host clones in governed `*_320dp_min_viewport_test.dart` (use `pumpDialogs320At`; showDialog-launcher hosts with `Builder` are not matched), inline `MaterialApp` hosts in panel/dialog/shell families listed in the checker, and private `_gameWithCapital` / `_gameWithPlayer` / `_gameWith` / `_buildGame` factories in `app_event_handler_debug_*_test.dart` (use `debug_handler_test_fixtures.dart`) (Refs #3730, #4035, #4048, #4058); rule `repo.app_test_no_duplicate_scaffolding` |
 | `tool/check_app_test_no_duplicate_golden_host.dart` | Full-tree gate: forbid duplicated golden-host scaffold clones outside `golden_capture_harness.dart` (Refs #3847, #4035); rule `repo.app_test_no_duplicate_golden_host` |
@@ -369,7 +369,7 @@ Large `app/test/**` widget/screen suites that are split into numbered fragments 
 
 **App e2e support file size (Refs #4075):** `repo.app_e2e_support_lib_file_size` caps every `packages/colonizethis_app_e2e_support/lib/**/*.dart` file at **700 physical lines**; `repo.app_e2e_support_test_file_size` caps every `packages/colonizethis_app_e2e_support/test/**/*.dart` file at **800 physical lines**. Oversized files must appear on the corresponding shrink-only allowlist; stale allowlist entries (missing path or file now ≤ cap) fail CI. Post-#4075 Slice A–C densify cleared both allowlists. Distinct from `repo.app_e2e_helper_location` (helper placement) and from repo-wide NCL 1000.
 
-**App test support LOC (Refs #4021, #4035, #4048, #4058, #4117):** `repo.app_test_support_loc` fails when physical LOC of all Dart under `app/test/support/` exceeds the documented post-densify ceiling (**8720**). Ratchet downward as densify slices shrink the tree.
+**App test support LOC (Refs #4021, #4035, #4048, #4058, #4117):** `repo.app_test_support_loc` fails when physical LOC of all Dart under `app/test/support/` exceeds the documented post-densify ceiling (**8302**). Ratchet downward as densify slices shrink the tree.
 
 **App test support file size (Refs #4048):** `repo.app_test_support_file_size` fails when any `app/test/support/**/*.dart` exceeds **600** physical lines unless listed on the shrink-only allowlist `appTestSupportFileSizeAllowlistForTests`. Stale allowlist entries (missing file, or file now ≤600) also fail.
