+++ app/test/ct_resource_cell_test.dart
@@ -3,40 +3,19 @@ import 'package:colonizethis_app/widgets/ct_resource_cell.dart';
 import 'package:colonizethis_app/widgets/ct_spacing.dart';
 import 'package:colonizethis_test/test.dart' show suppressLogsForTests;
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'support/app_shell_harness.dart';
+import 'ct_resource_cell_test_helpers.dart';
 
 void main() {
   suppressLogsForTests();
 
-  Future<void> pumpCell(WidgetTester tester, Widget child) async {
-    await tester.pumpWidget(
-      buildAppShell(
-        child: Scaffold(
-          body: Padding(
-            padding: const EdgeInsets.all(16),
-            child: SizedBox(width: 240, child: child),
-          ),
-        ),
-      ),
-    );
-    await tester.pump();
-  }
-
-  Widget tinyIcon(BuildContext context) => const SizedBox(
-        key: Key('test-icon'),
-        width: 20,
-        height: 20,
-      );
-
   group('CtResourceCell delta colour + prefix rules (R10)', () {
     testWidgets('positive delta renders +N in --success', (tester) async {
-      await pumpCell(
+      await pumpCtResourceCell(
         tester,
         CtResourceCell(
-          iconBuilder: tinyIcon,
+          iconBuilder: ctResourceCellTinyIcon,
           name: 'Grain',
           quantity: 1240,
           delta: 12,
@@ -48,10 +27,10 @@ void main() {
     });
 
     testWidgets('negative delta renders -N in --danger', (tester) async {
-      await pumpCell(
+      await pumpCtResourceCell(
         tester,
         CtResourceCell(
-          iconBuilder: tinyIcon,
+          iconBuilder: ctResourceCellTinyIcon,
           name: 'Timber',
           quantity: 920,
           delta: -40,
@@ -63,10 +42,10 @@ void main() {
     });
 
     testWidgets('zero delta renders 0 in --muted with no + prefix', (tester) async {
-      await pumpCell(
+      await pumpCtResourceCell(
         tester,
         CtResourceCell(
-          iconBuilder: tinyIcon,
+          iconBuilder: ctResourceCellTinyIcon,
           name: 'Meat',
           quantity: 870,
           delta: 0,
@@ -79,18 +58,16 @@ void main() {
     });
 
     testWidgets('null delta is not laid out (negative path)', (tester) async {
-      await pumpCell(
+      await pumpCtResourceCell(
         tester,
         CtResourceCell(
-          iconBuilder: tinyIcon,
+          iconBuilder: ctResourceCellTinyIcon,
           name: 'Coal',
           quantity: 170,
           // ignore: avoid_redundant_argument_values
           delta: null,
         ),
       );
-      // 170 is the quantity, not a delta. Verify that no extra Text widget is
-      // present beyond the name and quantity.
       final Iterable<Text> texts = tester
           .widgetList<Text>(find.descendant(
             of: find.byType(CtResourceCell),
@@ -133,10 +110,10 @@ void main() {
     testWidgets('renders icon + name + quantity in the documented order', (
       tester,
     ) async {
-      await pumpCell(
+      await pumpCtResourceCell(
         tester,
         CtResourceCell(
-          iconBuilder: tinyIcon,
+          iconBuilder: ctResourceCellTinyIcon,
           name: 'Grain',
           quantity: 1240,
         ),
@@ -156,10 +133,10 @@ void main() {
     testWidgets(
       'quantity always uses --accent-dim regardless of delta sign',
       (tester) async {
-        await pumpCell(
+        await pumpCtResourceCell(
           tester,
           CtResourceCell(
-            iconBuilder: tinyIcon,
+            iconBuilder: ctResourceCellTinyIcon,
             name: 'Iron',
             quantity: 430,
             delta: -20,
@@ -171,10 +148,10 @@ void main() {
     );
 
     testWidgets('name styles default to --fg body text', (tester) async {
-      await pumpCell(
+      await pumpCtResourceCell(
         tester,
         CtResourceCell(
-          iconBuilder: tinyIcon,
+          iconBuilder: ctResourceCellTinyIcon,
           name: 'Wool',
           quantity: 380,
         ),
@@ -184,24 +161,15 @@ void main() {
     });
 
     testWidgets('name truncates with ellipsis under tight width', (tester) async {
-      await tester.pumpWidget(
-        buildAppShell(
-          child: Scaffold(
-            body: Padding(
-              padding: const EdgeInsets.all(16),
-              child: SizedBox(
-                width: 80,
-                child: CtResourceCell(
-                  iconBuilder: tinyIcon,
-                  name: 'Refined Sugar — Long Label',
-                  quantity: 50,
-                ),
-              ),
-            ),
-          ),
+      await pumpCtResourceCellFixedWidth(
+        tester,
+        CtResourceCell(
+          iconBuilder: ctResourceCellTinyIcon,
+          name: 'Refined Sugar — Long Label',
+          quantity: 50,
         ),
+        width: 80,
       );
-      await tester.pump();
       final Text name = tester.widget<Text>(
         find.descendant(
           of: find.byType(CtResourceCell),
@@ -215,40 +183,13 @@ void main() {
   });
 
   group('CtResourceCell compact typography (Refs #2862 S9 / C9–C10)', () {
-    Future<void> pumpFixedWidth(
-      WidgetTester tester,
-      Widget cell, {
-      required double width,
-    }) async {
-      await tester.pumpWidget(
-        buildAppShell(
-          child: Scaffold(
-            body: Align(
-              alignment: Alignment.topLeft,
-              child: SizedBox(width: width, child: cell),
-            ),
-          ),
-        ),
-      );
-      await tester.pump();
-    }
-
-    RenderParagraph paragraphFor(WidgetTester tester, String text) {
-      return tester.renderObject<RenderParagraph>(
-        find.descendant(
-          of: find.byType(CtResourceCell),
-          matching: find.text(text),
-        ),
-      );
-    }
-
     testWidgets(
       'name + quantity render at the compact ~10 px mockup size; delta at ~9 px',
       (tester) async {
-        await pumpFixedWidth(
+        await pumpCtResourceCellFixedWidth(
           tester,
           CtResourceCell(
-            iconBuilder: tinyIcon,
+            iconBuilder: ctResourceCellTinyIcon,
             name: 'Grain',
             quantity: 1240,
             delta: 45,
@@ -262,7 +203,6 @@ void main() {
         expect(name.style?.fontSize, CtResourceCell.nameFontSize);
         expect(qty.style?.fontSize, CtResourceCell.quantityFontSize);
         expect(delta.style?.fontSize, CtResourceCell.deltaFontSize);
-        // The delta is one step smaller than the quantity per the mockup.
         expect(
           CtResourceCell.deltaFontSize,
           lessThan(CtResourceCell.quantityFontSize),
@@ -279,22 +219,19 @@ void main() {
           'Sugar Cane',
           'Refined Sugar',
         ]) {
-          await pumpFixedWidth(
+          await pumpCtResourceCellFixedWidth(
             tester,
             CtResourceCell(
-              iconBuilder: tinyIcon,
+              iconBuilder: ctResourceCellTinyIcon,
               name: name,
               quantity: 430,
               delta: -20,
             ),
-            // Wide enough that the compact name + intrinsic qty/delta cluster
-            // all fit; the legacy flex-share trailing cluster would have
-            // squeezed the name into ellipsis at this width.
             width: 320,
           );
 
           expect(
-            paragraphFor(tester, name).didExceedMaxLines,
+            ctResourceCellParagraphFor(tester, name).didExceedMaxLines,
             isFalse,
             reason: '"$name" must render in full without ellipsis clipping '
                 'per #2862 C9.',
@@ -309,18 +246,18 @@ void main() {
         const name = 'Iron';
         const width = 220.0;
 
-        await pumpFixedWidth(
+        await pumpCtResourceCellFixedWidth(
           tester,
-          CtResourceCell(iconBuilder: tinyIcon, name: name, quantity: 430),
+          CtResourceCell(iconBuilder: ctResourceCellTinyIcon, name: name, quantity: 430),
           width: width,
         );
         final noDeltaExceeded =
-            paragraphFor(tester, name).didExceedMaxLines;
+            ctResourceCellParagraphFor(tester, name).didExceedMaxLines;
 
-        await pumpFixedWidth(
+        await pumpCtResourceCellFixedWidth(
           tester,
           CtResourceCell(
-            iconBuilder: tinyIcon,
+            iconBuilder: ctResourceCellTinyIcon,
             name: name,
             quantity: 430,
             delta: -20,
@@ -328,10 +265,8 @@ void main() {
           width: width,
         );
         final withDeltaExceeded =
-            paragraphFor(tester, name).didExceedMaxLines;
+            ctResourceCellParagraphFor(tester, name).didExceedMaxLines;
 
-        // Name must not ellipsis whether or not a delta is present (C10 /
-        // #3999 inset-parity trailing is intrinsic; delta may reclaim flex).
         expect(noDeltaExceeded, isFalse);
         expect(withDeltaExceeded, isFalse);
       },
@@ -339,86 +274,21 @@ void main() {
   });
 
   group('CtResourceCell panel-wide amount alignment (#3999)', () {
-    /// Representative Available 3-column cell width (~800 dp wide Production
-    /// with side-by-side subpanels → ~120–130 dp slots).
-    const double gridCellWidth = 120;
-
-    Future<void> pumpFixedWidth(
-      WidgetTester tester,
-      Widget cell, {
-      required double width,
-    }) async {
-      await tester.pumpWidget(
-        buildAppShell(
-          child: Scaffold(
-            body: Align(
-              alignment: Alignment.topLeft,
-              child: SizedBox(width: width, child: cell),
-            ),
-          ),
-        ),
-      );
-      await tester.pump();
-    }
-
-    Future<void> pumpAlignedColumn(
-      WidgetTester tester,
-      List<Widget> cells, {
-      required double cellWidth,
-    }) async {
-      await tester.pumpWidget(
-        buildAppShell(
-          child: Scaffold(
-            body: Align(
-              alignment: Alignment.topLeft,
-              child: SizedBox(
-                width: cellWidth,
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  crossAxisAlignment: CrossAxisAlignment.stretch,
-                  children: cells,
-                ),
-              ),
-            ),
-          ),
-        ),
-      );
-      await tester.pump();
-    }
-
-    double quantityRight(WidgetTester tester, Finder cellFinder) {
-      final Finder qty = find.descendant(
-        of: cellFinder,
-        matching: find.byKey(CtResourceCell.quantityTextKey),
-      );
-      expect(qty, findsOneWidget);
-      return tester.getTopRight(qty).dx;
-    }
-
-    double quantityWidth(WidgetTester tester, Finder cellFinder) {
-      final Finder qty = find.descendant(
-        of: cellFinder,
-        matching: find.byKey(CtResourceCell.quantityTextKey),
-      );
-      expect(qty, findsOneWidget);
-      return tester.getSize(qty).width;
-    }
-
     testWidgets(
       'quantity 0 with null delta stays visible at grid width',
       (tester) async {
-        await pumpFixedWidth(
+        await pumpCtResourceCellFixedWidth(
           tester,
           CtResourceCell(
-            iconBuilder: tinyIcon,
+            iconBuilder: ctResourceCellTinyIcon,
             name: 'Grain',
             quantity: 0,
           ),
-          width: gridCellWidth,
+          width: kCtResourceCellGridCellWidth,
         );
         expect(find.text('0'), findsOneWidget);
         expect(
-          quantityWidth(tester, find.byType(CtResourceCell)),
+          ctResourceCellQuantityWidth(tester, find.byType(CtResourceCell)),
           greaterThan(1),
         );
       },
@@ -427,20 +297,20 @@ void main() {
     testWidgets(
       'quantity 0 with negative delta shows both values at grid width',
       (tester) async {
-        await pumpFixedWidth(
+        await pumpCtResourceCellFixedWidth(
           tester,
           CtResourceCell(
-            iconBuilder: tinyIcon,
+            iconBuilder: ctResourceCellTinyIcon,
             name: 'Grain',
             quantity: 0,
             delta: -16,
           ),
-          width: gridCellWidth,
+          width: kCtResourceCellGridCellWidth,
         );
         expect(find.text('0'), findsOneWidget);
         expect(find.text('-16'), findsOneWidget);
         expect(
-          quantityWidth(tester, find.byType(CtResourceCell)),
+          ctResourceCellQuantityWidth(tester, find.byType(CtResourceCell)),
           greaterThan(1),
         );
       },
@@ -449,15 +319,15 @@ void main() {
     testWidgets(
       'quantity 0 with positive delta shows both values at grid width',
       (tester) async {
-        await pumpFixedWidth(
+        await pumpCtResourceCellFixedWidth(
           tester,
           CtResourceCell(
-            iconBuilder: tinyIcon,
+            iconBuilder: ctResourceCellTinyIcon,
             name: 'Grain',
             quantity: 0,
             delta: 12,
           ),
-          width: gridCellWidth,
+          width: kCtResourceCellGridCellWidth,
         );
         expect(find.text('0'), findsOneWidget);
         expect(find.text('+12'), findsOneWidget);
@@ -470,33 +340,35 @@ void main() {
         const Key tinKey = ValueKey<String>('cell_tin');
         const Key sugarKey = ValueKey<String>('cell_sugar');
         const Key refinedKey = ValueKey<String>('cell_refined');
-        await pumpAlignedColumn(
+        await pumpCtResourceCellAlignedColumn(
           tester,
           <Widget>[
             CtResourceCell(
               key: tinKey,
-              iconBuilder: tinyIcon,
+              iconBuilder: ctResourceCellTinyIcon,
               name: 'Tin',
               quantity: 4,
             ),
             CtResourceCell(
               key: sugarKey,
-              iconBuilder: tinyIcon,
+              iconBuilder: ctResourceCellTinyIcon,
               name: 'Sugar Cane',
               quantity: 4,
             ),
             CtResourceCell(
               key: refinedKey,
-              iconBuilder: tinyIcon,
+              iconBuilder: ctResourceCellTinyIcon,
               name: 'Refined Sugar',
               quantity: 4,
             ),
           ],
-          cellWidth: gridCellWidth,
+          cellWidth: kCtResourceCellGridCellWidth,
         );
-        final double tinX = quantityRight(tester, find.byKey(tinKey));
-        final double sugarX = quantityRight(tester, find.byKey(sugarKey));
-        final double refinedX = quantityRight(tester, find.byKey(refinedKey));
+        final double tinX = ctResourceCellQuantityRight(tester, find.byKey(tinKey));
+        final double sugarX =
+            ctResourceCellQuantityRight(tester, find.byKey(sugarKey));
+        final double refinedX =
+            ctResourceCellQuantityRight(tester, find.byKey(refinedKey));
         expect(tinX, closeTo(sugarX, 0.5));
         expect(sugarX, closeTo(refinedX, 0.5));
       },
@@ -507,20 +379,19 @@ void main() {
       (tester) async {
         const Key nullDeltaKey = ValueKey<String>('inset_null');
         const Key withDeltaKey = ValueKey<String>('inset_delta');
-        // Representative wide Available 3-column slot (~800 dp Production).
         const double wideCellWidth = 240;
-        await pumpAlignedColumn(
+        await pumpCtResourceCellAlignedColumn(
           tester,
           <Widget>[
             CtResourceCell(
               key: nullDeltaKey,
-              iconBuilder: tinyIcon,
+              iconBuilder: ctResourceCellTinyIcon,
               name: 'Meat',
               quantity: 0,
             ),
             CtResourceCell(
               key: withDeltaKey,
-              iconBuilder: tinyIcon,
+              iconBuilder: ctResourceCellTinyIcon,
               name: 'Grain',
               quantity: 0,
               delta: -16,
@@ -564,18 +435,18 @@ void main() {
       (tester) async {
         const Key shortKey = ValueKey<String>('parity_short');
         const Key longKey = ValueKey<String>('parity_long');
-        await pumpAlignedColumn(
+        await pumpCtResourceCellAlignedColumn(
           tester,
           <Widget>[
             CtResourceCell(
               key: shortKey,
-              iconBuilder: tinyIcon,
+              iconBuilder: ctResourceCellTinyIcon,
               name: 'Tin',
               quantity: 4,
             ),
             CtResourceCell(
               key: longKey,
-              iconBuilder: tinyIcon,
+              iconBuilder: ctResourceCellTinyIcon,
               name: 'Refined Sugar',
               quantity: 4,
             ),
@@ -583,8 +454,8 @@ void main() {
           cellWidth: 240,
         );
         expect(
-          quantityRight(tester, find.byKey(shortKey)),
-          closeTo(quantityRight(tester, find.byKey(longKey)), 0.5),
+          ctResourceCellQuantityRight(tester, find.byKey(shortKey)),
+          closeTo(ctResourceCellQuantityRight(tester, find.byKey(longKey)), 0.5),
         );
       },
     );
@@ -594,18 +465,18 @@ void main() {
       (tester) async {
         const Key shortKey = ValueKey<String>('worker_short');
         const Key longKey = ValueKey<String>('worker_long');
-        await pumpAlignedColumn(
+        await pumpCtResourceCellAlignedColumn(
           tester,
           <Widget>[
             CtResourceCell(
               key: shortKey,
-              iconBuilder: tinyIcon,
+              iconBuilder: ctResourceCellTinyIcon,
               name: 'Masters',
               quantity: 6,
             ),
             CtResourceCell(
               key: longKey,
-              iconBuilder: tinyIcon,
+              iconBuilder: ctResourceCellTinyIcon,
               name: 'Journeymen',
               quantity: 6,
             ),
@@ -613,8 +484,8 @@ void main() {
           cellWidth: 160,
         );
         expect(
-          quantityRight(tester, find.byKey(shortKey)),
-          closeTo(quantityRight(tester, find.byKey(longKey)), 0.5),
+          ctResourceCellQuantityRight(tester, find.byKey(shortKey)),
+          closeTo(ctResourceCellQuantityRight(tester, find.byKey(longKey)), 0.5),
         );
       },
     );
@@ -622,40 +493,28 @@ void main() {
     testWidgets(
       'magnitude matrix A–H: visibility + alignment at grid width',
       (tester) async {
-        const List<({String id, int qty, int? delta, String qtyText})> cases =
-            <({String id, int qty, int? delta, String qtyText})>[
-          (id: 'A', qty: 0, delta: null, qtyText: '0'),
-          (id: 'B', qty: 0, delta: -16, qtyText: '0'),
-          (id: 'C', qty: 0, delta: 12, qtyText: '0'),
-          (id: 'D', qty: 4, delta: null, qtyText: '4'),
-          (id: 'E', qty: 999, delta: null, qtyText: '999'),
-          (id: 'F', qty: 9999, delta: null, qtyText: '9,999'),
-          (id: 'G', qty: 9999, delta: -999, qtyText: '9,999'),
-          (id: 'H', qty: -5, delta: null, qtyText: '-5'),
-        ];
-
-        for (final case_ in cases) {
+        for (final case_ in kCtResourceCellMagnitudeMatrixCases) {
           const Key refKey = ValueKey<String>('matrix_ref');
           final Key caseKey = ValueKey<String>('matrix_${case_.id}');
-          await pumpAlignedColumn(
+          await pumpCtResourceCellAlignedColumn(
             tester,
             <Widget>[
               CtResourceCell(
                 key: refKey,
-                iconBuilder: tinyIcon,
+                iconBuilder: ctResourceCellTinyIcon,
                 name: 'Tin',
                 quantity: case_.qty,
                 delta: case_.delta,
               ),
               CtResourceCell(
                 key: caseKey,
-                iconBuilder: tinyIcon,
+                iconBuilder: ctResourceCellTinyIcon,
                 name: 'Refined Sugar',
                 quantity: case_.qty,
                 delta: case_.delta,
               ),
             ],
-            cellWidth: gridCellWidth,
+            cellWidth: kCtResourceCellGridCellWidth,
           );
 
           final Finder caseCell = find.byKey(caseKey);
@@ -665,7 +524,7 @@ void main() {
             reason: 'Case ${case_.id}: quantity ${case_.qtyText} must be visible',
           );
           expect(
-            quantityWidth(tester, caseCell),
+            ctResourceCellQuantityWidth(tester, caseCell),
             greaterThan(1),
             reason: 'Case ${case_.id}: quantity must have non-zero layout width',
           );
@@ -679,8 +538,8 @@ void main() {
             );
           }
           expect(
-            quantityRight(tester, find.byKey(refKey)),
-            closeTo(quantityRight(tester, caseCell), 0.5),
+            ctResourceCellQuantityRight(tester, find.byKey(refKey)),
+            closeTo(ctResourceCellQuantityRight(tester, caseCell), 0.5),
             reason: 'Case ${case_.id}: quantity anchors must align',
           );
         }
@@ -691,21 +550,21 @@ void main() {
       'delta sits immediately to the right of quantity',
       (tester) async {
         const Key deltaKey = ValueKey<String>('adj_delta');
-        await pumpAlignedColumn(
+        await pumpCtResourceCellAlignedColumn(
           tester,
           <Widget>[
             CtResourceCell(
               key: deltaKey,
-              iconBuilder: tinyIcon,
+              iconBuilder: ctResourceCellTinyIcon,
               name: 'Wool',
               quantity: 4,
               delta: -40,
             ),
           ],
-          cellWidth: gridCellWidth,
+          cellWidth: kCtResourceCellGridCellWidth,
         );
         final Finder deltaCell = find.byKey(deltaKey);
-        final double qtyRight = quantityRight(tester, deltaCell);
+        final double qtyRight = ctResourceCellQuantityRight(tester, deltaCell);
         final double deltaLeft =
             tester.getTopLeft(find.descendant(
               of: deltaCell,
@@ -721,10 +580,10 @@ void main() {
     testWidgets(
       'painted trailing cluster right edge matches leading icon inset',
       (tester) async {
-        await pumpFixedWidth(
+        await pumpCtResourceCellFixedWidth(
           tester,
           CtResourceCell(
-            iconBuilder: tinyIcon,
+            iconBuilder: ctResourceCellTinyIcon,
             name: 'Timber',
             quantity: 920,
             delta: -40,
